pyproject.toml
[tool.poetry]
name = "parqu"
version = "1.0.0"
description = "Utility to dump parquet file metadata quickly"
authors = ["kenny.yeung <kenny.yeung@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8"
pyarrow = "^6.0.0"
docopt-ng = "^0.7.2"
logzero = "^1.7.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.11b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
